Replace AssertJAssertions parser stubs with 'assertj-core' classpath reference.

+++ src/test/kotlin/org/openrewrite/java/testing/assertj/JUnitAssertArrayEqualsToAssertThatTest.kt
@@ -115,7 +115,7 @@ class JUnitAssertArrayEqualsToAssertThatTest : JavaRecipeTest {
 
                 @Test
                 public void test() {
-                    float eps = .2d;
+                    double eps = .2d;
                     assertArrayEquals(new double[]{1.0d, 2.0d, 3.0d}, notification(), eps);
                 }
                 private double[] notification() {
@@ -133,7 +133,7 @@ class JUnitAssertArrayEqualsToAssertThatTest : JavaRecipeTest {
 
                 @Test
                 public void test() {
-                    float eps = .2d;
+                    double eps = .2d;
                     assertThat(notification()).containsExactly(new double[]{1.0d, 2.0d, 3.0d}, within(eps));
                 }
                 private double[] notification() {
